requirements.txt
appdirs==1.4.4
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
idna==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
prometheus-client==0.11.0
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
six==1.16.0
tomli==1.2.0
urllib3==1.26.6
